backend/requirements.txt
fastapi
uvicorn
py-algorand-sdk
google-generativeai
python-dotenv
requests
firebase-admin
apscheduler
yfinance
